cap9-cli: Rearrange fetch module.
