+++ kernel-ewasm/cap9-cli/src/fetch/mod.rs
@@ -1,0 +1,42 @@
+/// Module for the fetch interface.
+
+use web3::futures::Future;
+use web3::contract::{Contract, Options};
+use web3::types::{Address, U256, H256};
+// use web3::types::TransactionReceipt;
+use web3::Transport;
+use rustc_hex::FromHex;
+use rustc_hex::ToHex;
+// use ethabi::Token::Uint;
+use crate::conn;
+use crate::conn::EthConn;
+use crate::project::LocalProject;
+use cap9_std::proc_table::cap::*;
+use pwasm_abi;
+use std::fs::File;
+use std::fmt;
+use cap9_std::proc_table::ProcPointer;
+use cap9_std::proc_table;
+use cap9_core::*;
+use cap9_core::Error;
+use cap9_core::Read;
+use crate::constants;
+use crate::deploy::{from_common_u256, to_common_u256, to_common_h256,
+    from_common_address, to_common_address
+};
+use std::collections::{HashMap, HashSet};
+
+mod acl;
+pub use acl::*;
+
+mod kernel;
+pub use kernel::*;
+
+mod map;
+pub use map::*;
+
+mod procedure;
+pub use procedure::*;
+
+mod utils;
+pub use utils::*;
